+++ src/lib/utils/functions.cpp
@@ -161,9 +161,9 @@ void yield_for(std::chrono::microseconds us)
 
 pid_t detachProcess(const bool closeDescriptors)
 {
-	GK_LOG_FUNC
+	//GK_LOG_FUNC
 
-	GKLog(trace, "detaching process")
+	//GKLog(trace, "detaching process")
 
 	pid_t pid;
 
@@ -173,20 +173,20 @@ pid_t detachProcess(const bool closeDescriptors)
 
 	// parent exit
 	if(pid > 0) {
-		GKLog2(trace, "first fork done. pid : ", pid)
+		//GKLog2(trace, "first fork done. pid : ", pid)
 		std::this_thread::sleep_for(std::chrono::milliseconds(20));
-		GKLog(trace, "exiting parent")
+		//GKLog(trace, "exiting parent")
 		exit(EXIT_SUCCESS);
 	}
-	else {
-		GKLog(trace, "continue child execution")
-	}
+	//else {
+	//	GKLog(trace, "continue child execution")
+	//}
 
 	// new session for child process
 	if(setsid() == -1)
 		throw GLogiKExcept("session creation failure");
 
-	GKLog(trace, "new session done")
+	//GKLog(trace, "new session done")
 
 	// Ignore signals
 	std::signal(SIGCHLD, SIG_IGN);
@@ -198,14 +198,14 @@ pid_t detachProcess(const bool closeDescriptors)
 
 	// parent exit
 	if(pid > 0) {
-		GKLog2(trace, "second fork done. pid : ", pid)
+		//GKLog2(trace, "second fork done. pid : ", pid)
 		std::this_thread::sleep_for(std::chrono::milliseconds(20));
-		GKLog(trace, "exiting parent")
+		//GKLog(trace, "exiting parent")
 		exit(EXIT_SUCCESS);
 	}
-	else {
-		GKLog(trace, "continue child execution")
-	}
+	//else {
+	//	GKLog(trace, "continue child execution")
+	//}
 
 	umask(S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP);
 	if(chdir("/") == -1)
@@ -224,11 +224,11 @@ pid_t detachProcess(const bool closeDescriptors)
 		stdout = std::fopen("/dev/null", "w+");
 		stderr = std::fopen("/dev/null", "w+");
 
-		GKLog(trace, "descriptors closed, process daemonized")
+		//GKLog(trace, "descriptors closed, process daemonized")
 	}
 
 	pid = getpid();
-	GKLog2(trace, "returning pid : ", pid)
+	//GKLog2(trace, "returning pid : ", pid)
 	return pid;
 }